+++ CHANGELOG.md
@@ -12,14 +12,22 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### Changes
 
+### Fixes
+
+### Breaks
+
+
+## 1.35.1 - (2020-06-11)
+---
+
+### Changes
+
 - [#2094](https://gitlab.com/meltano/meltano/-/issues/2094) Consistently refer to plugin types and names in CLI output
 
 ### Fixes
 
 - [#2092](https://gitlab.com/meltano/meltano/-/issues/2092) Only install plugins related to plugins of the specified type when running `meltano install <plugin_type> --include-related`
 - [#2093](https://gitlab.com/meltano/meltano/-/issues/2093) Print error when transform plugin is installed before dbt
-
-### Breaks
 
 
 ## 1.35.0 - (2020-06-09)